refactor: move strategies to own folder
