+++ api/Documents/Strategies/FileStrategy.cs
@@ -2,6 +2,7 @@ using System.IO;
 using System.Text;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.WebUtilities;
+using Scv.Api.Documents.Strategies;
 using Scv.Api.Models.Document;
 using Scv.Api.Services.Files;
