+++ api/Documents/Strategies/ReportStrategy.cs
@@ -1,5 +1,6 @@
 using System.IO;
 using System.Threading.Tasks;
+using Scv.Api.Documents.Strategies;
 using Scv.Api.Models.CourtList;
 using Scv.Api.Models.Document;
 using Scv.Api.Services;
